DRY-ing things up
