+++ apps/eternal-devotion/src/app/utils/index.ts
@@ -3,3 +3,4 @@ import classnames from './classnames';
 export { classnames };
 export * from './routing';
 export * from './image_preloader';
+export * from './responsive_image';
